SDK cleanup
Clean up the wasmer-sdk crate

* Set version
  (should not track the main runtime version)
* Add some docs
* Move package publishing to submodule
* Rename PublishError to PackagePublishError
* Run expensive operations on blocking threads
