+++ lib/sdk/README.md
@@ -6,3 +6,8 @@ including functionality for:
 * Building and publishing packages
 * Deploying apps
 * ...
+
+## Architecture Support
+
+This crate currently only works on native architectures, and does not support
+browser Webassembly yet. Support will likely added in the future.
