+++ lib/sdk/src/app/mod.rs
@@ -1,2 +1,4 @@
+//! Provides tooling related to apps deployed on Wasmer Edge.
+
 pub mod deploy;
 pub mod deploy_remote_build;
