+++ lib/sdk/src/lib.rs
@@ -1,2 +1,4 @@
+//! wasmer-sdk is a utility library for interacting with the Wasmer ecosystem.
+
 pub mod app;
 pub mod package;
